pyproject.toml
[project]
name = "swa-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pyside6>=6.9.0",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "ultralytics>=8.3.149",
]
